+++ src/components/Input/__docs__/File.stories.tsx
@@ -1,9 +1,18 @@
+import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { File } from '../components';
+import { File as FileComponent, FileProps } from '../components';
+
+const Component = (props: FileProps) => {
+  const [value, setValue] = useState<File | null>(null);
+
+  return (
+    <FileComponent value={value} onChange={setValue} {...props} />
+  );
+};
 
 const meta = {
   title: 'Example/Input/File',
-  component: File,
+  component: Component,
   tags: ['autodocs'],
   parameters: {
     layout: 'centered',
@@ -14,7 +23,7 @@ const meta = {
       control: 'text',
     },
   },
-} satisfies Meta<typeof File>;
+} satisfies Meta<typeof FileComponent>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;