+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.0.4 - 2024-10-12
+### Added
+- Add loading prop to Button.Link component
+
+### Changed
+- Improve Button loader styles
+
 ## 1.0.3 - 2024-10-08
 ### Added
 - Add onClick handler to Button.Link component